requirements.txt
langchain
langchain-openai
langgraph
docker
streamlit
python-dotenv